requirements.txt
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
coverage==7.4.4
cryptography==42.0.5
exceptiongroup==1.2.0
fastapi==0.110.1
flake8==7.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
mccabe==0.7.0
packaging==24.0
pluggy==1.4.0
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
pyflakes==3.2.0
pytest==8.1.1
python-dotenv==1.0.1
PyYAML==6.0.1
redis==5.0.3
sniffio==1.3.1
starlette==0.37.2
tomli==2.0.1
typing_extensions==4.11.0
uvicorn==0.29.0
watchfiles==0.21.0
websockets==12.0
